extract plugin-specific code into a separate module
